Work on startup settings
- Move CLRStartup to common location.
- Add documentation about CLR_SETTINGS fields.
